clean up clinical service worker thread setup
